Move PanelManager#constructMessageData to ManagedChannel#constructMessageData for clarity
